Rename: docs 설정 파일 이름 변경
